requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.0
Pillow>=10.4.0
python-dateutil==2.8.2
cryptography>=42.0.0
email-validator==2.1.0
